fix(providers): tkms HTTPS-only + extract shared decodeEntities
CodeRabbit findings on #1555:

- Major (security): tkms.mjs's resolveConfig() accepted http: origins even
  though the API fetch uses redirect: 'error' — an HTTP origin would send the
  request in cleartext and hard-fail on any HTTPS redirect. Now HTTPS-only,
  with a regression test.
- Trivial: decodeEntities()/NAMED_ENTITIES was duplicated verbatim across
  deutschebahn.mjs and hecklerkoch.mjs. Extracted to providers/_html-entities.mjs
  (mirroring the same extraction just done on #1639), including the range +
  surrogate guard both copies already had.

The max_pages/resolveMaxPages test-coverage finding (test-all.mjs:11763) was
stale — that coverage already exists in tests/providers/deutschebahn.test.mjs
from the #1500 test-suite restructure; no action needed there.
